refactor: VideoController의 역할 분리
